Read from stdin and stdout in separate threads
The shell command used to block when it produced large amounts
of output on process-error. The parent read process-output and
process-error sequentially. When the output on process-error
excedeed the pipe buffer, the shell command would hang, waiting
for the pipe to drain.
NB: This patch will only work with a threads enabled sbcl.

darcs-hash:20071106205112-fe858-29606aadb80616154887a171ae467ce5c33e22eb.gz